refactor: folder cleaning
