+++ src/app/examples/product-test/page.tsx
@@ -1,5 +1,10 @@
 import ProductList from "@/components/modules/product/List";
+import { Suspense } from "react";
 
 export default function page() {
-  return <ProductList />;
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <ProductList />
+    </Suspense>
+  );
 }
