+++ src/app/examples/subject-test/page.tsx
@@ -1,5 +1,10 @@
 import SubjectList from "@/components/modules/subject/List";
+import { Suspense } from "react";
 
 export default function page() {
-  return <SubjectList />;
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <SubjectList />
+    </Suspense>
+  );
 }
