+++ src/app/examples/teacher-test/page.tsx
@@ -1,6 +1,11 @@
 import TeacherList from "@/components/modules/teacher/List";
+import { Suspense } from "react";
 
 
 export default function page() {
-  return <TeacherList />;
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <TeacherList />
+    </Suspense>
+  );
 }
