+++ src/components/modules/teacher/Form.tsx
@@ -1,370 +1,283 @@
 'use client';
 
+import { getAllInstitutions } from '@/backend/actions/institution.action';
 import { ActionResult } from '@/backend/actions/IServerAction';
-import { createTeacher } from '@/backend/actions/teacher.action';
-import { TeacherFormValues } from '@/schemas/teacher';
 import {
-  Button,
-  Checkbox,
-  Chip,
-  DatePicker,
-  Input,
-  Select,
-  SelectItem,
-  Textarea,
-} from '@heroui/react';
-import { CalendarDate } from '@internationalized/date';
+  createTeacher,
+  getTeacherById,
+  getTeacherDesignations,
+  updateTeacher,
+} from '@/backend/actions/teacher.action';
+import { FormCheckbox } from '@/components/forms/FormCheckbox';
+import { FormProvider } from '@/components/forms/FormContainer';
+import { FormDatePicker } from '@/components/forms/FormDatePicker';
+import { FormInput } from '@/components/forms/FormInput';
+import { FormSelect } from '@/components/forms/FormSelect';
+import { useFormData } from '@/hooks/useFormData';
+import { useFormOptions } from '@/hooks/useFormOptions';
+import { teacherFormSchema, TeacherFormValues, TeacherWithUser } from '@/schemas/teacher';
+import { Spinner } from '@heroui/react';
+import { zodResolver } from '@hookform/resolvers/zod';
 import { Teacher } from '@prisma/client';
-import { useActionState, useEffect } from 'react';
+import { useActionState, useEffect, useMemo } from 'react';
+import { useForm } from 'react-hook-form';
 
-// Define the teacher data type to match the schema
-export type TeacherData = {
-  firstName: string;
-  lastName: string;
-  email: string;
-  phone: string;
-  institutionId: string;
-  designation: string;
-  pdsId: string | null;
-  joiningDate: Date | null;
-  address: string | null;
-  district: string | null;
-  specialization: string | null;
-  status: boolean;
-  deleted: boolean;
-};
+export type DrawerMode = 'create' | 'read' | 'edit';
 
 interface TeacherFormProps {
-  defaultValues: Partial<TeacherFormValues>;
-  institutions: { id: string; name: string }[];
-  designations: string[];
+  teacherId?: string;
+  mode: DrawerMode;
   isReadOnly?: boolean;
-  onSubmit?: (data: TeacherFormValues) => void;
+  onSuccess?: () => void;
 }
 
-export default function TeacherForm({
-  defaultValues,
-  institutions,
-  designations,
-  isReadOnly = false,
-  onSubmit,
-}: TeacherFormProps) {
-  // Initialize action state with correct type
-  const initialState: ActionResult<Teacher> = {
+export function TeacherForm({ teacherId, mode, isReadOnly = false, onSuccess }: TeacherFormProps) {
+  // Initialize action state
+  const initialState: ActionResult<TeacherWithUser> = {
     success: false,
     errors: [],
   };
 
-  // Set up the action state with the correct type
-  const [state, formAction, isPending] = useActionState(createTeacher, initialState);
+  // Get data for editing
+  const { data: teacherData, isLoading: isLoadingData } = useFormData<TeacherWithUser>(
+    // async (id: string) => {
+    //   const result = await getTeacherById(id);
+    //   if (!result.success || !result.data) {
+    //     return { success: false, errors: [{ field: 'root', message: 'Failed to load teacher' }] };
+    //   }
+    //   // Cast the result to TeacherWithUser since we know it includes user and institution data
+    //   return { success: true, data: result.data as unknown as TeacherWithUser };
+    // },
+    getTeacherById,
+    teacherId,
+    mode !== 'create'
+  );
+  console.log('teacherData>>', teacherData);
+  
+
+  // Set up React Hook Form with Zod resolver
+  const methods = useForm<TeacherFormValues>({
+    resolver: zodResolver(teacherFormSchema),
+    mode: 'onChange',
+    defaultValues: {
+      firstName: '',
+      lastName: '',
+      email: '',
+      phone: '',
+      institutionId: '',
+      designation: '',
+      pdsId: '',
+      joiningDate: null,
+      address: '',
+      district: '',
+      specialization: '',
+      status: true,
+      userId: '',
+    },
+  });
+
+  const formValues = methods.watch();
+
+  // Configuration for institutions, levels, and designations
+  const fieldConfig = useMemo(
+    () => ({
+      institutions: {
+        fetchFunction: () => getAllInstitutions(),
+        dependencies: [],
+      },
+      designations: {
+        fetchFunction: () => getTeacherDesignations(),
+        dependencies: [],
+      },
+    }),
+    []
+  );
+
+  // Use the hook with this configuration
+  const { options, loading } = useFormOptions(fieldConfig, formValues);
+
+  // Set up action states for create and update
+  const [createState, createAction, isCreatePending] = useActionState(createTeacher, initialState);
+
+  // For update mode with validation
+  const wrappedUpdateTeacher = (state: ActionResult<Teacher>, formData: FormData) => {
+
+    
+    if (teacherId) {
+      return updateTeacher(state, teacherId, formData);
+    }
+    return {
+      success: false as const,
+      errors: [{ field: 'root', message: 'Missing teacher ID' }],
+    };
+  };
+
+  const [updateState, updateAction, isUpdatePending] = useActionState(
+    (state: ActionResult<Teacher>, formData: FormData) => updateTeacher(state, teacherId!, formData),
+    initialState
+  );
+
+  // Determine which state and action to use based on mode
+  const state = mode === 'edit' ? updateState : createState;
+  const formAction = mode === 'edit' ? updateAction : createAction;
+  const isPending = isUpdatePending || isCreatePending;
+
+  // Update form values when teacherData changes
+  useEffect(() => {
+    if (teacherData && Object.keys(teacherData).length > 0) {
+      methods.reset({
+        firstName: teacherData.firstName || '',
+        lastName: teacherData.lastName || '',
+        email: teacherData.email || '',
+        phone: teacherData.phone || '',
+        institutionId: teacherData.institutionId || '',
+        designation: teacherData.designation || '',
+        pdsId: teacherData.pdsId || '',
+        joiningDate: teacherData.joiningDate
+          ? new Date(teacherData.joiningDate).toISOString().split('T')[0]
+          : null,
+        address: teacherData.address || '',
+        district: teacherData.district || '',
+        specialization: teacherData.specialization || '',
+        status: teacherData.status !== false,
+        userId: teacherData.userId || '',
+      });
+
+      // methods.setValue('userId', teacherData.userId || '');
+    }
+  }, [teacherData, methods.reset]);
+  
 
   // Handle success state
   useEffect(() => {
-    if (state.success && onSubmit) {
-      onSubmit(defaultValues as TeacherFormValues);
+    if (state.success && onSuccess) {
+      onSuccess();
     }
-  }, [state.success, onSubmit, defaultValues]);
+  }, [state.success, onSuccess]);
 
-  // Extract field errors into a more usable format
-  const getFieldErrors = (fieldName: string): string[] | undefined => {
-    if (!state.success && state.errors) {
-      if (state.errors.length > 0) {
-        console.log('state errors>>', state.errors);
-      }
+  // Show loading spinner while initial data is loading
+  if (isLoadingData) {
+    return (
+      <div className="flex justify-center items-center h-64">
+        <Spinner color="primary" size="lg" />
+      </div>
+    );
+  }
 
-      const fieldErrors = state.errors
-        .filter((error) => error.field === fieldName)
-        .map((error) => error.message);
+  // Format options for select components
+  const institutionOptions =
+    options?.institutions?.map((institution) => ({
+      id: institution.id,
+      value: institution.id,
+      label: institution.name,
+    })) || [];
 
-      return fieldErrors.length > 0 ? fieldErrors : undefined;
-    }
-    return undefined;
-  };
-
-  // Get all form-level errors (non-field specific)
-  const formErrors =
-    !state.success && state.errors
-      ? state.errors.filter(
-          (error) =>
-            error.field === 'root' || error.field === 'unknown' || typeof error.field === 'number'
-        )
-      : [];
-
-  // Parse date for DatePicker
-  const parseDateValue = (dateValue: any): CalendarDate | null => {
-    if (!dateValue) return null;
-
-    try {
-      // If dateValue is already a CalendarDate, return it
-      if (dateValue instanceof CalendarDate) {
-        return dateValue;
-      }
-
-      // Handle string values (ISO date strings)
-      if (typeof dateValue === 'string') {
-        // Parse ISO date string to native Date object first
-        const date = new Date(dateValue);
-
-        // Check if the date is valid
-        if (isNaN(date.getTime())) {
-          return null;
-        }
-
-        // Convert Date to CalendarDate
-        return new CalendarDate(
-          date.getFullYear(),
-          date.getMonth() + 1, // JavaScript months are 0-based
-          date.getDate()
-        );
-      }
-
-      // Handle JavaScript Date objects
-      if (dateValue instanceof Date) {
-        return new CalendarDate(
-          dateValue.getFullYear(),
-          dateValue.getMonth() + 1, // JavaScript months are 0-based
-          dateValue.getDate()
-        );
-      }
-
-      return null;
-    } catch (e) {
-      console.error('Error parsing date:', e);
-      return null;
-    }
-  };
-
-  // Convert empty strings to null for optional fields
-  const prepareDefaultValue = (value: any): string => {
-    return value === null || value === undefined ? '' : String(value);
-  };
+  const designationOptions =
+    options?.designations?.map((designation) => ({
+      id: designation.id,
+      value: designation.id,
+      label: designation.name,
+    })) || [];
 
   return (
-    <form action={formAction} className="space-y-6">
-      {/* Hidden teacherId field for edit mode */}
-      <input type="hidden" name="teacherId" value={defaultValues.teacherId || ''} />
+    <FormProvider<TeacherFormValues>
+      methods={methods}
+      actionMethod={formAction}
+      isReadOnly={isReadOnly}
+      isPending={isPending}
+      serverErrors={state.errors}
+      showSuccessMessage={state.success}
+      successMessage={`Teacher successfully ${mode === 'edit' ? 'updated' : 'created'}!`}
+      submitText={mode === 'edit' ? 'Update Teacher' : 'Create Teacher'}
+    >
+      <input type="hidden" name="userId" value={teacherData?.userId} />
+      <FormInput
+        name="firstName"
+        label="First Name"
+        placeholder="Enter first name"
+        isRequired
+        isDisabled={isReadOnly || isPending}
+      />
 
-      {/* Form-level errors */}
-      {formErrors.length > 0 && (
-        <div className="p-3 mb-4 text-sm text-white bg-red-500 rounded-md">
-          {formErrors.map((error, index) => (
-            <div key={index}>{error.message}</div>
-          ))}
-        </div>
-      )}
+      <FormInput
+        name="lastName"
+        label="Last Name"
+        placeholder="Enter last name"
+        isRequired
+        isDisabled={isReadOnly || isPending}
+      />
 
-      {/* First Name */}
-      <div>
-        <Input
-          name="firstName"
-          label="First Name"
-          placeholder="Enter first name"
-          defaultValue={prepareDefaultValue(defaultValues.firstName)}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('firstName')}
-          errorMessage={getFieldErrors('firstName')?.join(', ')}
-          className="w-full"
-          isRequired
-        />
-      </div>
+      <FormInput
+        name="email"
+        label="Email"
+        type="email"
+        placeholder="Enter email address"
+        isRequired
+        isDisabled={isReadOnly || isPending}
+      />
 
-      {/* Last Name */}
-      <div>
-        <Input
-          name="lastName"
-          label="Last Name"
-          placeholder="Enter last name"
-          defaultValue={prepareDefaultValue(defaultValues.lastName)}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('lastName')}
-          errorMessage={getFieldErrors('lastName')?.join(', ')}
-          className="w-full"
-          isRequired
-        />
-      </div>
+      <FormInput
+        name="phone"
+        label="Phone"
+        placeholder="Enter phone number"
+        isDisabled={isReadOnly || isPending}
+      />
 
-      {/* Email */}
-      <div>
-        <Input
-          name="email"
-          type="email"
-          label="Email"
-          placeholder="Enter email address"
-          defaultValue={prepareDefaultValue(defaultValues.email)}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('email')}
-          errorMessage={getFieldErrors('email')?.join(', ')}
-          className="w-full"
-          isRequired
-        />
-      </div>
+      <FormSelect
+        name="institutionId"
+        label="Institution"
+        options={institutionOptions}
+        placeholder={loading?.institutions ? 'Loading institutions...' : 'Select institution'}
+        isRequired
+        isDisabled={isReadOnly || isPending || loading?.institutions}
+        isLoading={loading?.institutions}
+      />
 
-      {/* Phone */}
-      <div>
-        <Input
-          name="phone"
-          type="tel"
-          label="Phone"
-          placeholder="Enter phone number"
-          defaultValue={prepareDefaultValue(defaultValues.phone)}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('phone')}
-          errorMessage={getFieldErrors('phone')?.join(', ')}
-          className="w-full"
-        />
-      </div>
+      <FormSelect
+        name="designation"
+        label="Designation"
+        options={designationOptions}
+        placeholder={loading?.designations ? 'Loading designations...' : 'Select designation'}
+        isRequired
+        isDisabled={isReadOnly || isPending || loading?.designations}
+        isLoading={loading?.designations}
+      />
 
-      {/* Institution */}
-      <div>
-        <Select
-          name="institutionId"
-          label="Institution"
-          placeholder="Select institution"
-          defaultSelectedKeys={defaultValues.institutionId ? [defaultValues.institutionId] : []}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('institutionId')}
-          errorMessage={getFieldErrors('institutionId')?.join(', ')}
-          className="w-full"
-          isRequired
-        >
-          {institutions.map((institution) => (
-            <SelectItem key={institution.id} textValue={institution.id}>
-              {institution.name}
-            </SelectItem>
-          ))}
-        </Select>
-      </div>
+      <FormInput
+        name="pdsId"
+        label="PDS ID"
+        placeholder="Enter PDS ID"
+        isDisabled={isReadOnly || isPending}
+      />
 
-      {/* Designation */}
-      <div>
-        <Select
-          name="designation"
-          label="Designation"
-          placeholder="Select designation"
-          defaultSelectedKeys={defaultValues.designation ? [defaultValues.designation] : []}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('designation')}
-          errorMessage={getFieldErrors('designation')?.join(', ')}
-          className="w-full"
-          isRequired
-        >
-          {designations.map((designation) => (
-            <SelectItem key={designation} textValue={designation}>
-              {designation}
-            </SelectItem>
-          ))}
-        </Select>
-      </div>
+      <FormDatePicker
+        name="joiningDate"
+        label="Joining Date"
+        isDisabled={isReadOnly || isPending}
+      />
 
-      {/* PDS ID */}
-      <div>
-        <Input
-          name="pdsId"
-          label="PDS ID"
-          placeholder="Enter PDS ID (optional)"
-          defaultValue={prepareDefaultValue(defaultValues.pdsId)}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('pdsId')}
-          errorMessage={getFieldErrors('pdsId')?.join(', ')}
-          className="w-full"
-        />
-      </div>
+      <FormInput
+        name="address"
+        label="Address"
+        placeholder="Enter address"
+        isDisabled={isReadOnly || isPending}
+      />
 
-      {/* Joining Date */}
-      <div>
-        <DatePicker
-          name="joiningDate"
-          label="Joining Date"
-          aria-label="Select joining date"
-          // value={parseDateValue(defaultValues.joiningDate)}
-          isDisabled={isReadOnly || isPending}
-          className="w-full"
-        />
-        {getFieldErrors('joiningDate') && (
-          <div className="mt-1 text-xs text-red-500">
-            {getFieldErrors('joiningDate')?.join(', ')}
-          </div>
-        )}
-      </div>
+      <FormInput
+        name="district"
+        label="District"
+        placeholder="Enter district"
+        isDisabled={isReadOnly || isPending}
+      />
 
-      {/* District */}
-      <div>
-        <Select
-          name="district"
-          label="District"
-          placeholder="Select district (optional)"
-          defaultSelectedKeys={defaultValues.district ? [defaultValues.district] : []}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('district')}
-          errorMessage={getFieldErrors('district')?.join(', ')}
-          className="w-full"
-        >
-          {institutions.map((institution) => (
-            <SelectItem key={institution.id} textValue={institution.id}>
-              {institution.name}
-            </SelectItem>
-          ))}
-        </Select>
-      </div>
+      <FormInput
+        name="specialization"
+        label="Specialization"
+        placeholder="Enter specialization"
+        isDisabled={isReadOnly || isPending}
+      />
 
-      {/* Address */}
-      <div>
-        <Textarea
-          name="address"
-          label="Address"
-          placeholder="Enter address (optional)"
-          defaultValue={prepareDefaultValue(defaultValues.address)}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('address')}
-          errorMessage={getFieldErrors('address')?.join(', ')}
-          className="w-full"
-        />
-      </div>
-
-      {/* Specialization */}
-      <div>
-        <Input
-          name="specialization"
-          label="Specialization"
-          placeholder="Enter specialization (optional)"
-          defaultValue={prepareDefaultValue(defaultValues.specialization)}
-          isDisabled={isReadOnly || isPending}
-          isInvalid={!!getFieldErrors('specialization')}
-          errorMessage={getFieldErrors('specialization')?.join(', ')}
-          className="w-full"
-        />
-      </div>
-
-      {/* Status */}
-      <div>
-        <Checkbox
-          name="status"
-          defaultSelected={defaultValues.status !== false}
-          isDisabled={isReadOnly || isPending}
-          value={'on'}
-        >
-          Active Status
-        </Checkbox>
-        {getFieldErrors('status') && (
-          <div className="mt-1 text-xs text-red-500">{getFieldErrors('status')?.join(', ')}</div>
-        )}
-      </div>
-
-      {/* Submit button */}
-      <div className="flex justify-end">
-        <Button type="submit" color="primary" isLoading={isPending} isDisabled={isPending}>
-          {defaultValues.teacherId ? 'Update Teacher' : 'Create Teacher'}
-        </Button>
-      </div>
-
-      {/* Success message */}
-      {state.success && (
-        <div className="p-3 mt-4 text-green-700 bg-green-100 rounded-md">
-          <Chip color="success">
-            Teacher successfully {defaultValues.teacherId ? 'updated' : 'created'}!
-          </Chip>
-        </div>
-      )}
-    </form>
+      <FormCheckbox name="status" label="Active Status" isDisabled={isReadOnly || isPending} />
+    </FormProvider>
   );
 }
